[AI] implementing opencode in my workflow
This is the age of AI. Opencode is a TUI tool that easily integrates in
NeoVIM and vscode.

Until now I was switching between vscode and neovim due to tools. But if
I can use this tool the editor is not that important anymore as I have a
standard tool.